Move the plugins to a specialized subpackage similarly to Avocado VT
This is also slightly improving the plugin module structure by
moving all entry points in a subfolder (with the exception of
a few new ones that are still improved upstream).
