+++ setup.py
@@ -33,19 +33,20 @@ setup(name='avocado-framework-plugins-i2n',
       author='Intra2net AG',
       author_email='support@intra2net.com',
       url='http://github.com/intra2net/avocado-i2n',
-      packages=['avocado_i2n', 'avocado_i2n.cartgraph', 'avocado_i2n.vmnet'],
+      packages=['avocado_i2n', 'avocado_i2n.plugins', 'avocado_i2n.cartgraph', 'avocado_i2n.vmnet'],
       package_data={'avocado_i2n.vmnet': ['templates/*.template']},
       install_requires=['avocado-framework-plugins-vt', 'aexpect'],
       data_files=[(get_dir(['etc', 'avocado', 'conf.d']),
                    ['etc/avocado/conf.d/i2n.conf'])],
       entry_points={
-         'avocado.plugins.cli': [
-              'auto = avocado_i2n.auto:Auto',
+          'avocado.plugins.cli': [
+              'auto = avocado_i2n.plugins.auto:Auto',
               ],
           'avocado.plugins.cli.cmd': [
-              'manu = avocado_i2n.manu:Manu',
+              'manu = avocado_i2n.plugins.manu:Manu',
               ],
           'avocado.plugins.runner': [
+              # TODO: wait for the upstream to make the default loader a plugin as well
               'traverser = avocado_i2n.runner:CartesianRunner',
               ],
           },
